requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-levenshtein  # or Levenshtein, but pip levenshtein
python-dotenv
tomli  # Fallback for older Python if needed, but 3.11 has tomllib
